pyproject.toml
[tool.poetry]
name = "flask-sample"
version = "0.1.0"
description = ""
authors = ["hH"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0"
Flask-SQLAlchemy = "^3.1.1"
flask-marshmallow = "^1.2.1"
Flask-Security = "^5.5.2"
Flask-Script = "^2.0.6"
Flask-Migrate = "^4.0.7"
Flask-Admin = "^1.6.1"
flask-restx = "^1.3.0"
marshmallow-sqlalchemy = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
